% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{align_rasters}
\alias{align_rasters}
\title{Align Rasters of possibly differnt diminsions/stats}
\usage{
align_rasters(huc6, name.dir)
}
\arguments{
\item{huc6}{A HUC6 unit}

\item{name.dir}{the directory where data is, and, will be written}
}
\value{
on disk raster files (tiffs)
}
\description{
Align catchmask and HAND rasters to catchmask grid
}
