% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_tiles.R
\name{clip_it}
\alias{clip_it}
\title{Clip the HAND raster}
\source{
Attribution: The \code{elevatr} R package.
\url{https://github.com/jhollist/elevatr}
}
\usage{
clip_it(rast, loc, expand, clip)
}
\description{
Clip the HAND raster
}
\keyword{internal}
