% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_gen.R
\name{compute_flat_tub}
\alias{compute_flat_tub}
\title{Get flat tub}
\usage{
compute_flat_tub(stage, df)
}
\arguments{
\item{stage}{Stage value}

\item{df}{HAND/Slope dataframe. See \link{get_src}.}
}
\description{
Get flat tub
}
\keyword{internal}
