% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crop_project}
\alias{crop_project}
\title{Crop and Reproject tiff using GDAL}
\usage{
crop_project(input, output, name, aoi.path, method)
}
\arguments{
\item{input}{a path to the input file}

\item{output}{the path to the output file}

\item{name}{the name of the region used in LivingFlood}

\item{aoi.path}{a path to an AOI object on disk}

\item{method}{resampling method (e.g. 'near')}
}
\value{
on disk rasters files (tiffs)
}
\description{
Wrapper around gdalwarp. Projects rasters to 3857
}
