% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_tiles.R
\name{download_tiles}
\alias{download_tiles}
\title{Get HAND Tiles from an AWS S3 Bucket}
\usage{
download_tiles(locations, z = 0, prj, expand = NULL, ...)
}
\arguments{
\item{z}{The zoom level to return. The HAND tiles only use \code{z = 0}.}

\item{prj}{PROJ.4 string for input bbox}

\item{expand}{A numeric value of a distance, in map units, used to expand the
bounding box that is used to fetch the tiles. This can
be used for features that fall close to the edge of a tile and
additional area around the feature is desired. Default is NULL.}

\item{...}{Extra configuration parameters to be passed to httr::GET.  Common
usage is to adjust timeout.  This is done as
\code{config=timeout(x)} where \code{x} is a numeric value in
seconds. Multiple configuration functions may be passed as a
vector.}

\item{bbx}{a \code{sf::st_bbox} object that is used to select x,y,z tiles.}
}
\description{
This function uses AWS S3 to retrieve HAND rasters from the geotiff folder.
It accepts a \code{sf::st_bbox} object as input and returns a single raster
object covering that extent.
}
\keyword{internal}
