% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ut.R
\name{download_ut}
\alias{download_ut}
\title{Download HAND, RC, and Cathmask rasters by HUC6}
\usage{
download_ut(HUC6, outdir, type = "all")
}
\arguments{
\item{HUC6}{A HUC6 unit to download}

\item{outdir}{The folder path where data should be downloaded and extracted}

\item{type}{The type of data to download. Options include 'catchmask'. 'hand', or 'rating curve'.
If 'all' is used (DEFAULT) all three type are downloaded.}
}
\description{
This function downloads staged HAND data to a local directory.
}
