% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ut.R
\name{downloader}
\alias{downloader}
\title{Function to download data from URL to out directory using httr.}
\usage{
downloader(dir, url)
}
\arguments{
\item{dir}{path to output directory}

\item{url}{the location of the online resource}
}
\description{
General downloader
}
