% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_download_raw.R
\name{downloadRaw}
\alias{downloadRaw}
\title{Downlaod RAW hand data}
\usage{
downloadRaw(base.path, HUC6, type)
}
\arguments{
\item{base.path}{the path to main Living Flood directory}

\item{HUC6}{the HUC6 of interest}

\item{type}{the type of data desired. Acceptable values are: 'hand', 'catchmask', 'rating'}
}
\value{
the requested file downloaded as a GeoTIFF for RDA file
}
\description{
Function to isolate HUC6 regions of interest and download the appropriate data from the University of Texas server,
}
\examples{
\dontrun{
downloadRaw(base.path = base.path, HUC6 = 180702, type = 'hand')
downloadRaw(base.path = base.path, HUC6 = 180702, type = 'catchmask')
downloadRaw(base.path = base.path, HUC6 = 180702, type = 'rating')
}
}
\author{
Mike Johnson
}
\keyword{internal}
