% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_tiles.R
\name{estimate_raster_size}
\alias{estimate_raster_size}
\title{Estimate download size of DEMs}
\source{
Attribution: The \code{elevatr} R package.
\url{https://github.com/jhollist/elevatr}
}
\usage{
estimate_raster_size(locations, z = NULL)
}
\arguments{
\item{locations}{the locations}

\item{z}{zoom level if source is aws}
}
\description{
Estimate download size of DEMs
}
\keyword{internal}
