% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_gen.R
\name{find_comids}
\alias{find_comids}
\title{Find COMIDs for an AOI via the NHD}
\usage{
find_comids(aoi)
}
\arguments{
\item{aoi}{An area of interest. Can be retrieved via \code{AOI::aoi_get}.}
}
\value{
A \code{factor vector} of COMIDs.
}
\description{
Find COMIDs for an AOI via the NHD
}
\keyword{internal}
