% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findHUC6}
\alias{findHUC6}
\title{Identify HUC6 Units within AOI}
\usage{
findHUC6(AOI, level = 6)
}
\arguments{
\item{AOI}{an AOI}

\item{level}{HUC level to extract (default = 6)}
}
\value{
a sf object
}
\description{
Identify HUC6 Units within AOI
}
