% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_findNHD.R
\name{findNHD}
\alias{findNHD}
\title{Find National Hydrography Data Stream Networks}
\usage{
findNHD(AOI = NULL, ids = FALSE)
}
\arguments{
\item{AOI}{If TRUE, returns a list of COMIDS for NHD reaches}

\item{ids}{If TRUE, data is written to a HydroData folder in users working directory.}
}
\description{
\code{findNHD} returns a list of \code{Spatial*} Objects cropped to an Area of Interest.\cr\cr
}
\author{
Mike Johnson
}
