% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_tiles.R
\name{get_hand_xy}
\alias{get_hand_xy}
\title{Get HAND Tile XY Coordinates}
\usage{
get_hand_xy(aoi)
}
\arguments{
\item{aoi}{Area of interest}
}
\description{
Get HAND Tile XY Coordinates
}
\keyword{internal}
