% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rc_table}
\alias{get_rc_table}
\title{Extract Rating Curve Table}
\usage{
get_rc_table(HUC6, raw.dir, comids)
}
\arguments{
\item{HUC6}{a HUC6 unit}

\item{raw.dir}{the directory where RAW downloaded HAND data goes}

\item{comids}{a vector of COMIDs of interest}
}
\value{
a rating curve tabel
}
\description{
Extract Rating Curve Table
}
