% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_gen.R
\name{get_roughness}
\alias{get_roughness}
\title{Get roughness value prediction}
\usage{
get_roughness(pathlength, arbolatesu, lengthkm, areasqkm, slope)
}
\arguments{
\item{pathlength}{The distance from the flowline outlet to the end of
the network along the main network path}

\item{arbolatesu}{Arbolate sum: the cumulative length of the upstream
drainage network (mainstem and tributaries) from
the outlet of the catchment}

\item{lengthkm}{Length of a flowline in kilometers}

\item{areasqkm}{Drainage area in km^2}

\item{slope}{The longitudinal slope.}
}
\value{
A roughness value
}
\description{
This function querys a REST API serving the Gradient Boosting machine
learning model developed in the in-review paper:

\strong{Johnson, J.M., Eyelade D., Clarke K.C, Singh-Mohudpur, J. (2021)
\emph{“Characterizing Reach-level Empirical Roughness Along the National
Hydrography Network: Developing DEM-based Synthetic Rating Curves.”}}
}
