% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src_gen.R
\name{get_src}
\alias{get_src}
\title{Generate Synthetic Rating Curve}
\usage{
get_src(
  comids,
  hand,
  stage = 0:20,
  progress = TRUE,
  slope_scale = 111120,
  verbose = getOption("verbose")
)
}
\arguments{
\item{comids}{A set of COMIDs. See \link{find_comids}.}

\item{hand}{A HAND \code{raster}.}

\item{stage}{A set of stage values. Defaults to \code{0:20}.}

\item{progress}{If \code{TRUE}, output a progress bar.}

\item{slope_scale}{Ratio of vertical units to horizontal.
Uses \code{slope_scale = 111120} by default.
See \href{https://gdal.org/programs/gdaldem.html}{gdaldem}.}

\item{verbose}{Verbose output for HAND download.}
}
\value{
A \code{data.frame} representing the SRC table.
}
\description{
Generate Synthetic Rating Curve
}
