% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get HAND data for an AOI}
\usage{
getData(AOI = NULL, name = NULL, read.path = NULL,
  write.path = NULL)
}
\arguments{
\item{AOI}{a spatial geometery to subset with}

\item{name}{the name of the region your are studying (default will be RegionXXX)}

\item{read.path}{a path to look for existing HAND data}

\item{write.path}{the path to write main LIVINGFLOOD directory. Default will be the users working directory.}
}
\value{
the path the the directory storing processed data
}
\description{
Download raw HAND, catchment and rating curve file and crop to an area of interest
}
\author{
Mike Johnson
}
