% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawData.R
\name{getRawData}
\alias{getRawData}
\title{Download, process, and save HAND, catchmask and Rating Curve Data}
\usage{
getRawData(AOI, dir, name)
}
\arguments{
\item{AOI}{a AOI spatial object to map floods within}

\item{dir}{a main working directory for all floodmapping projects}

\item{name}{a name for current project}
}
\value{
a list of file paths
}
\description{
Download, process, and save HAND, catchmask and Rating Curve Data
}
