% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivingFlood.R
\docType{package}
\name{LivingFlood}
\alias{LivingFlood}
\alias{LivingFlood-package}
\title{LivingFlood}
\description{
\code{LivingFlood} package
}
\details{
Generate near real-time and forecasted flood extents using the National Water Model

See the README on github
}
