% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_tiles.R
\name{loc_check}
\alias{loc_check}
\title{Function to check input type and projection.}
\usage{
loc_check(locations, prj = NULL)
}
\arguments{
\item{locations}{AOI}

\item{prj}{CRS}
}
\value{
\code{sf} object
}
\description{
Function to check input type and projection.
}
\keyword{internal}
