% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Map flood extents from processed data}
\usage{
map(name.dir, write = TRUE, add = 0)
}
\arguments{
\item{name.dir}{the directory where processed data is stored}

\item{write}{logical. Should output rasters be written to disk (default = TRUE)}

\item{add}{numeric. A uniform depth to add to simulated inudation}
}
\value{
a raster stack of innudation maps
}
\description{
This function takes a folder of processed data and generates a HAND inudation raster for each timestep
}
\author{
Mike Johnson
}
