% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2.R
\name{map_flood}
\alias{map_flood}
\title{Map flood extents from processed data}
\usage{
map_flood(hand.path, catchment.path, flows.path, add = 0, threshold = 0)
}
\arguments{
\item{hand.path}{a path to HAND tif}

\item{catchment.path}{a path to catchmask tif}

\item{flows.path}{a path to a flows nc file}

\item{add}{numeric. A uniform depth to add to simulated inudation}

\item{threshold}{a value to exclude values less then}
}
\value{
a raster stack of innudation maps
}
\description{
This function takes a folder of processed data and generates a HAND inundation raster for each timestep
}
