% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_tiles.R
\name{merge_hand_tiles}
\alias{merge_hand_tiles}
\title{Merge Rasters}
\usage{
merge_hand_tiles(
  raster_list,
  target_prj,
  method = "bilinear",
  return_raster = TRUE
)
}
\arguments{
\item{raster_list}{a list of raster file paths to be mosaiced}

\item{target_prj}{the target projection of the output raster}

\item{method}{the method for resampling/reprojecting. Default is 'bilinear'.
Options can be found on the GDAL documentation site
for \code{gdalwarp}.}

\item{returnRaster}{if TRUE, return a raster object (default),
else, return the file path to the object}
}
\description{
Merge multiple downloaded raster files into a single file.
The input \code{target_prj} describes the projection for the new grid.
}
\keyword{internal}
