% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_rasters}
\alias{merge_rasters}
\title{Merge Rasters}
\usage{
merge_rasters(name.dir, name)
}
\arguments{
\item{name.dir}{a directory send the files}

\item{name}{a name of a study unit}
}
\value{
a list of local HAND and catchmask filepaths
}
\description{
Merge rasters from different HUC6 units
}
