% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_metaAOI.R
\name{metaAOI}
\alias{metaAOI}
\title{Get AOI meta data}
\usage{
metaAOI(AOI)
}
\arguments{
\item{AOI}{a spatial, raster or sf object}
}
\value{
a list of features describing the AOI
}
\description{
Break a spatial geometry into descriptive meta data
}
\examples{
\dontrun{
AOI = getAOI('Colorado Springs') \%>\% metaAOI
}
}
\author{
Mike Johnson
}
\keyword{internal}
