% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_mosaic.R
\name{mosaic.lf}
\alias{mosaic.lf}
\title{Crop and merge and write raster sets}
\usage{
mosaic.lf(input = NULL, AOI = NULL)
}
\arguments{
\item{input}{a raster stack}

\item{bb}{a bounding box to crop to}

\item{write.path}{a path to write processed data}
}
\description{
Crop and merge datasets that cross HUC6 boundaries
}
\author{
Mike Johnson
}
\keyword{internal}
