% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2.R
\name{nc_flow_mat}
\alias{nc_flow_mat}
\title{Build named flow matrix from CONUS netcdf file}
\usage{
nc_flow_mat(x, comids)
}
\arguments{
\item{x}{a file path to a flow NetCDF}

\item{comids}{a vector of COMIDs}
}
\value{
data.frame
}
\description{
Build named flow matrix from CONUS netcdf file
}
