% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_processData.R
\name{processData}
\alias{processData}
\title{Process Raw HAND Data to an AOI}
\usage{
processData(AOI = NULL, raw.dir = NULL, write.path = NULL)
}
\arguments{
\item{AOI}{a list containing spatial a spatial geometry and a collection of NHD reaches}

\item{raw.dir}{the path where the raw HAND data is stored}

\item{write.path}{the path to write the processed data to}
}
\description{
Crop, merge and filter raw HUC6 HAND products to an AOI
}
\author{
Mike Johnson
}
\keyword{internal}
