% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_tiles.R
\name{clip_it}
\alias{clip_it}
\title{Clip the HAND raster}
\usage{
clip_it(rast, loc, expand, clip)
}
\arguments{
\item{rast}{\code{raster} object}

\item{loc}{AOI to clip \code{rast} to}

\item{expand}{Amount to expand (in meters)}

\item{clip}{Clipping type}
}
\value{
Clipped \code{raster}
}
\description{
Clip the HAND raster
}
\keyword{internal}
