% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_tiles.R
\name{estimate_raster_size}
\alias{estimate_raster_size}
\title{Estimate download size of DEMs}
\usage{
estimate_raster_size(locations)
}
\arguments{
\item{locations}{the locations}
}
\value{
\code{numeric} of estimated download size
}
\description{
Estimate download size of DEMs
}
\keyword{internal}
