% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_findNHD.R
\name{findNHD}
\alias{findNHD}
\title{Find National Hydrography Stream Networks}
\usage{
findNHD(AOI = NULL)
}
\arguments{
\item{AOI}{a spatial geometry to clip to}
}
\value{
a list with AOI geometry and NHD feature set
}
\description{
\code{findNHD} returns a list of \code{SimpleFeatures} object cropped to an Area of Interest.
}
\author{
Mike Johnson
}
\keyword{internal}
