% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFlows.R
\name{getFlows}
\alias{getFlows}
\title{Download streamflow forecasts}
\usage{
getFlows(name.dir, config = "medium_range", startDate = NULL,
  t = NULL, f = NULL, n = NULL)
}
\arguments{
\item{name.dir}{the directory with processed geospatial data}

\item{config}{the model configureation (default is 'medium_range')}

\item{startDate}{the date of interest given as "YYYYMMDD" defaults to current date}

\item{t}{time of forecast}

\item{f}{hours foward from t}

\item{n}{if f is null n can be used to limit the number of files returned}
}
\value{
a path to the processed data directory
}
\description{
Wrapper around the NWM package to download National Water Model data from the HydroShare THREDDS server
}
\author{
Mike Johnson
}
