% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_tiles.R
\name{loc_check}
\alias{loc_check}
\title{function to check input type and projection.  All input types convert to a
SpatialPointsDataFrame for point elevation and bbx for raster.}
\source{
Attribution: The \code{elevatr} R package.
\url{https://github.com/jhollist/elevatr}
}
\usage{
loc_check(locations, prj = NULL)
}
\description{
function to check input type and projection.  All input types convert to a
SpatialPointsDataFrame for point elevation and bbx for raster.
}
\keyword{internal}
