% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hand_tiles.R
\name{proj_expand}
\alias{proj_expand}
\title{Project bounding box and Expand if needed}
\usage{
proj_expand(locations, prj, expand)
}
\description{
Project bounding box and Expand if needed
}
\keyword{internal}
